+++ koa-test/src/contexts/projects.js
@@ -1,6 +1,6 @@
 import Router from 'koa-tree-router'
 import mongoose from 'mongoose'
-import validateGod from '../auth'
+import authorize from '../auth'
 
 ////////////
 // Schema //
@@ -29,7 +29,7 @@ router.get('/projects', async (ctx) => {
 })
 
 router.post('/projects', async (ctx) => {
-  validateGod(ctx)
+  authorize(ctx)
 
   const project = new Project(ctx.request.body)
   await project.save()
@@ -44,7 +44,7 @@ router.get('/projects/:id', async (ctx) => {
 })
 
 router.put('/projects/:id', async (ctx) => {
-  validateGod(ctx)
+  authorize(ctx)
 
   const project = await Project.findByIdAndUpdate(ctx.params.id, ctx.request.body)
 
@@ -52,7 +52,7 @@ router.put('/projects/:id', async (ctx) => {
 })
 
 router.delete('/projects/:id', async (ctx) => {
-  validateGod(ctx)
+  authorize(ctx)
 
   await Project.findByIdAndRemove(ctx.params.id)
